Simplified 'Blog' to 'Projects'
Changed 'Blog' to Project's because the more-complicated "blog" style was too much to try to keep up with (since it included category pages, tag pages, etc., and all of it was static). Switched to much more manageable "Projects" page, which will ONLY have my project breakdowns. Blog will be separate if I want to have one.